requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asgiref==3.6.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
backports.cached-property==1.0.2
beautifulsoup4==4.11.2
blessed==1.19.1
braceexpand==0.1.7
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.7
croniter==1.3.8
cryptography==39.0.0
cycler==0.11.0
Cython==0.29.33
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.2.3
Django==3.2.16
django-auditlog==2.2.2
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
dnspython==2.3.0
docopt==0.6.2
drf-excel==2.2.0
drf-flex-fields==1.0.0
einops==0.3.2
email-validator==1.3.1
et-xmlfile==1.1.0
fastapi==0.88.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
hmmlearn==0.2.8
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.12.0
hydra-core==1.3.1
HyperPyYAML==1.1.0
idna==3.4
importlib-metadata==6.0.0
inflect==6.0.2
inquirer==3.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
julius==0.2.7
kafka-python==2.0.2
kiwisolver==1.4.4
ldap3==2.9.1
librosa==0.9.2
lightning==1.9.0
lightning-cloud==0.5.19
lightning-utilities==0.6.0.post0
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.4
NEMO==4.4.1
nemo-toolkit==1.14.0
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.0
openpyxl==3.1.0
optuna==3.1.0
ordered-set==4.1.0
orjson==3.8.5
packaging==22.0
pandas==1.5.3
Pillow==9.4.0
pooch==1.6.0
primePy==1.3
protobuf==3.20.1
psutil==5.9.4
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
pyDeprecate==0.3.2
Pygments==2.14.0
PyJWT==2.6.0
pymodbus==3.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-editor==1.0.4
python-multipart==0.0.5
pytorch-lightning==1.6.5
pytorch-metric-learning==1.7.3
pytz==2022.7.1
PyYAML==6.0
readchar==4.0.3
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.1
scipy==1.10.0
semver==2.13.0
sentencepiece==0.1.97
shellingham==1.5.0.post1
simplejson==3.18.1
singledispatchmethod==1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
speechbrain==0.5.13
SQLAlchemy==2.0.0
sqlparse==0.4.3
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
text-unidecode==1.3
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.2
torchaudio==0.13.1
torchmetrics==0.11.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.0
typer==0.7.0
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.6
webdataset==0.2.31
websocket-client==1.5.0
websockets==10.4
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
yarl==1.8.2
youtokentome==1.0.6
zipp==3.12.0
